streamdata-generator/requirements.txt
google-cloud-bigquery
google-cloud-pubsub
google-cloud-bigquery-storage  
Faker
ratelimit
pandas==2.1.3